bionicpro-auth/requirements.txt
fastapi
uvicorn
python-keycloak
requests
python-multipart
jinja2
